% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterConcepts.R
\name{clusterConcepts}
\alias{clusterConcepts}
\title{clusterConcepts}
\usage{
clusterConcepts(dbconnection, cdmDatabaseSchema = NULL, method = "kmeans",
  clusterSize = 10, topicSize = NULL, scale = T,
  covariatesToInclude = NULL, indications = T, dayStart = 1,
  dayEnd = 30, use_min_obs = TRUE, min_obs = 100,
  extraparameters = NULL, updateProgress = NULL, ...)
}
\arguments{
\item{dbconnection}{using DatabaseConnector - connect to cdm database}

\item{cdmDatabaseSchema}{- cdm schema used to extract data from}

\item{method}{class:character - method used to do clustering (currently only supports kmeans)}

\item{clusterSize}{class:numeric - number of clusters returned,}

\item{topicSize}{class:numeric - number of topics in glrm}

\item{scale}{class:boolean - whether to use ingredience percentage scale for clustering}

\item{covariatesToInclude}{class:character vector - features to include: default NULL}

\item{indications}{class:boolean   extract drug indicator features;Default TRUE}

\item{dayStart}{class:integer number of days relative to condition_concept_code to start looking for drugs}

\item{dayEnd}{class:integer number of days relative to condition_concept_code to stop looking for drugs}

\item{use_min_obs}{class:boolean whether to remove ingredient features that are rare}

\item{min_obs}{clkass:integer threshold used when use_min_obs is TRUE to determine what is rare}
}
\value{
list contining definition data.frame containing columes for concept_id, covariate (cluster id)
}
\description{
Create topics by clusting condition_concept_ids based on ingredience counts
}
\examples{
clusterConcepts()
}
\keyword{OHDSI,}
\keyword{clustering}

