% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPeopleLatest.R
\name{clusterPeople}
\alias{clusterPeople}
\title{runs kmeans on the data}
\usage{
clusterPeople(clusterData, ageSpan = c(0, 100), gender = 8507,
  method = "kmeans", clusterSize = 10, normalise = T,
  covariatesToInclude = NULL, covariatesToExclude = NULL,
  covariatesGroups = NULL, loc = loc)
}
\arguments{
\item{ageSpan}{class:numeric - 1=0-20 year olds; 2=20-40 year olds; 3=40-60 year olds; 4=60-80 year olds;5=80-100 year olds}

\item{gender}{class:numeric - gender concept_id (8507- male; 8532-female)}

\item{method}{class:character - method used to do clustering (currently only supports kmeans)}

\item{clusterSize}{class:numeric - number of clusters returned,}

\item{normalise}{class:boolean - whether to center the data prior to clustering}

\item{covariatesToInclude}{class:character vector - features to include: default NULL}

\item{covariatesToExclude}{class:character  vector - features to exclude;Default NULL}

\item{covariatesGroups}{class:covariatecluster   result of clusterCovariate();Default NULL}

\item{loc}{class:character - directory where the results of the clustering are saved}

\item{cohortid}{class:numeric - id of cohort in cohort table}
}
\value{
...
}
\description{
runs kmeans on the data
}
\examples{
clusterRun()
}
\seealso{
DatabaseConnector, OhdsiRTools, SqlRender, ggplot2, reshape2, dplyr, plyr
}
\keyword{OHDSI,}
\keyword{clustering}

