% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPredictionModels.R
\name{developPredictionModels}
\alias{developPredictionModels}
\title{Develop prediction model to predict cluster probabilities}
\usage{
developPredictionModels(covariates, cluster)
}
\arguments{
\item{covariates}{An ffdf containing the covariates and values}

\item{cluster}{A dataframe containing row_ids and cluster columns}
}
\value{
A list containing:
\item{predictionModels}{ A named list containing the trained models for each cluster (the names are the clusters predicted by the models)}
}
\description{
This function creates a binary logistic regression model for each cluster
}
\details{
This function trains a predictive model using the input covariates to predict cluster membership
}
\keyword{OHDSI,}
\keyword{clustering}

