% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtract.R
\name{loadClusterData}
\alias{loadClusterData}
\title{Load the cluster data from a folder}
\usage{
loadClusterData(file, readOnly = FALSE)
}
\arguments{
\item{file}{The name of the folder containing the data.}

\item{readOnly}{If true, the data is opened read only.}
}
\value{
An object of class clusterData
}
\description{
\code{loadClusterData} loads an object of type \code{clusterData} from a folder in the file system.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}
\examples{
# todo

}

