% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runMain.R
\name{runMain}
\alias{runMain}
\title{cluster patients in your OHDSi cohort based on their medical history}
\usage{
runMain(dbconnection, cdmDatabaseSchema, workDatabaseSchema, outputFolder,
  cohortid = 100, agegroup = 3, gender = 8507, type = "history",
  method = "kmeans", clusterSize = 10, centerVal = T,
  covariatesToInclude = NULL, covariatesToExclude = NULL)
}
\arguments{
\item{dbconnection}{class:connectionDetails - the database connection details requires Library(DatabaseConnector)}

\item{cdmDatabaseSchema}{class:character - database schema containing database}

\item{workDatabaseSchema}{class:character -database you writing tables to}

\item{outputFolder}{class:character - directory output will be written to}

\item{cohortid}{class:numeric - id of cohort in cohort table}

\item{agegroup}{class:numeric - 1=0-20 year olds; 2=20-40 year olds; 3=40-60 year olds; 4=60-80 year olds;5=80-100 year olds}

\item{gender}{class:numeric - gender concept_id (8507- male; 8532-female)}

\item{type}{class:character - features used by clustering (currently only history),}

\item{method}{class:character - method used to do clustering (currently only supports kmeans)}

\item{clusterSize}{class:numeric - number of clusters returned,}

\item{centerVal}{class:boolean - whether to center the data prior to clustering}

\item{covariatesToInclude}{class:character vector - features to include: default NULL}

\item{covariatesToExclude}{class:character  vector - features to exclude;Default NULL}
}
\value{
NULL (writes results to output folder)
}
\description{
cluster patients in your OHDSi cohort based on their medical history
}
\examples{
runMain()
}
\seealso{
DatabaseConnector, OhdsiRTools, SqlRender, ggplot2, reshape2, dplyr, plyr
}
\keyword{OHDSI,}
\keyword{clustering}

