% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtract.R
\name{saveClusterData}
\alias{saveClusterData}
\title{Save the clustering data to folder}
\usage{
saveClusterData(cData, file)
}
\arguments{
\item{cData}{An object of type \code{clusterData} as generated using \code{dataExtract}.}

\item{file}{The name of the folder where the data will be written. The folder should not yet
exist.}
}
\description{
\code{saveClusterData} saves an object of type clusterData to folder.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}
\examples{
# todo

}

