% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterEval.R
\name{topicEval}
\alias{topicEval}
\title{topicEval}
\usage{
topicEval(clust.res, threshold = NULL)
}
\arguments{
\item{clust.res}{A list return by running clusterConcepts()}
}
\value{
A list containing:
\item{topicsOrdered}{ A data frame containing the mean of each feature per cluster}
\item{topicsMax}{ A data frame containing the standard deviation of each feature value per cluster}
\item{topicsKmeans}{ A data frame containing the fraction of each cluster with non-zero values for the feature}
}
\description{
This function simply calculates summaries of each topic and returns these
as a list.
}
\details{
This function only has one input, the clusterResults obtained by applying clusterConcepts
}
\keyword{OHDSI,}
\keyword{clustering}

