% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterEval.R
\name{clusterEval}
\alias{clusterEval}
\title{Summarises the differences between the clusters}
\usage{
clusterEval(clusterResult)
}
\arguments{
\item{clusterResult}{...}
}
\description{
Summarises the differences between the clusters
}
\examples{
clusterEval()

}
\keyword{OHDSI,}
\keyword{clustering}

