% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clusterEval.R
\name{clusterEval}
\alias{clusterEval}
\title{Summarises the differences between the clusters}
\usage{
clusterEval(cohortid, agegroup, gender, type)
}
\arguments{
\item{cohortid}{class:numeric - id of cohort in cohort table}

\item{agegroup}{class:numeric - 1=0-20 year olds; 2=20-40 year olds; 3=40-60 year olds; 4=60-80 year olds;5=80-100 year olds}

\item{gender}{class:numeric - gender concept_id (8507- male; 8532-female)}

\item{type}{class:character - features used by clustering (currently only history),}
}
\description{
Summarises the differences between the clusters
}
\examples{
clusterEval()
}
\keyword{OHDSI,}
\keyword{clustering}

