% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterVisual.R
\name{clusterVisual}
\alias{clusterVisual}
\title{Plots the different cluster visulisations}
\usage{
clusterVisual(clusterResult, saveLoc = getwd())
}
\arguments{
\item{saveLoc}{class:character - directory where the results of the clustering are saved}

\item{clusterresult}{output from applying clusterPeople()}
}
\description{
Plots the different cluster visulisations
}
\examples{
clusterVisual()

}
\keyword{OHDSI,}
\keyword{clustering}

