% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterVisual.R
\name{clusterVisual}
\alias{clusterVisual}
\title{Plots the different cluster visulisations}
\usage{
clusterVisual(clusterResult, saveLoc = getwd())
}
\arguments{
\item{saveLoc}{class:character - directory where the results of the clustering are saved}

\item{clusterresult}{output from applying clusterPeople()}
}
\description{
Plots barcharts of each cluster's center and saves to the directory specified by the user
}
\details{
This function only has two inputs, the clusterResults obtained by applying clusterPeople
and the location to save the plots.
}
\examples{
clusterVisual(clusterResult, saveLoc='C:/Documents')

}
\keyword{OHDSI,}
\keyword{clustering}

