% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataExtract.R
\name{dataExtract}
\alias{dataExtract}
\title{This extracts the history features for each person in the cohort with the specific age/gender}
\usage{
dataExtract(dbconnection, cdmDatabaseSchema, workDatabaseSchema,
  cohortid = 100, agegroup = 3, gender = 8507, type = "history")
}
\arguments{
\item{dbconnection}{class:connectionDetails - the database connection details requires Library(DatabaseConnector)}

\item{cdmDatabaseSchema}{class:character - database schema containing database}

\item{workDatabaseSchema}{class:character -database you writing tables to}

\item{cohortid}{class:numeric - id of cohort in cohort table}

\item{agegroup}{class:numeric - 1=0-20 year olds; 2=20-40 year olds; 3=40-60 year olds; 4=60-80 year olds;5=80-100 year olds}

\item{gender}{class:numeric - gender concept_id (8507- male; 8532-female)}

\item{type}{class:character - features used by clustering (currently only history),}
}
\value{
NULL (writes results to output folder)
}
\description{
This extracts the history features for each person in the cohort with the specific age/gender
}
\examples{
dataExtract()
}
\seealso{
DatabaseConnector, OhdsiRTools, SqlRender, ggplot2, reshape2, dplyr, plyr
}
\keyword{OHDSI,}
\keyword{clustering}

