% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtract.R
\name{dataExtract}
\alias{dataExtract}
\title{This extracts the history features for each person in the cohort with the specific age/gender}
\usage{
dataExtract(dbconnection = NULL, cdmDatabaseSchema = NULL,
  cohortDatabaseSchema = NULL, workDatabaseSchema = NULL, cohortid = 100,
  agegroup = NULL, gender = NULL, type = "condition",
  groupDef = "default", historyStart = 1, historyEnd = 180, loc = NULL)
}
\arguments{
\item{dbconnection}{class:connectionDetails - the database connection details requires Library(DatabaseConnector)}

\item{cdmDatabaseSchema}{class:character - database schema containing cdm tables}

\item{cohortDatabaseSchema}{class:character - database schema containing cohort}

\item{workDatabaseSchema}{class:character -database you writing tables to}

\item{cohortid}{class:numeric - id of cohort in cohort table}

\item{agegroup}{class:numeric - 1=0-20 year olds; 2=20-40 year olds; 3=40-60 year olds; 4=60-80 year olds;5=80-100 year olds}

\item{gender}{class:numeric - gender concept_id (8507- male; 8532-female)}

\item{type}{class:character - features used by clustering (currently only history),}

\item{historyStart}{....}

\item{historyEnd}{....}

\item{loc}{class:character - directory where the results of the clustering are saved}

\item{sqlType}{...}
}
\value{
clusterData class:clusterData - a list containing: strata- an ffdf containing the age/gender/row_id for each person in the cohort,
                                                           covariates - an ffdf containing the covariates for each person in the cohort
                                                           covariateRef - an ffdf containing details about the covariates
                                                           metadata - a list containing details about the data extraction
}
\description{
This extracts the history features for each person in the cohort with the specific age/gender
}
\examples{
dataExtract(gender=8507)
}
\seealso{
DatabaseConnector, OhdsiRTools, SqlRender, ggplot2, reshape2, dplyr, plyr
}
\keyword{OHDSI,}
\keyword{clustering}

