% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtract.R
\name{dataExtract}
\alias{dataExtract}
\title{This extracts the history features for each person in the cohort with the specific age/gender}
\usage{
dataExtract(dbconnection = NULL, cdmDatabaseSchema = NULL,
  cohortDatabaseSchema = NULL, workDatabaseSchema = NULL, cohortid = 100,
  agegroup = NULL, gender = NULL, type = "condition",
  groupDef = "default", historyStart = 1, historyEnd = 180, loc = NULL)
}
\arguments{
\item{dbconnection}{class:connectionDetails - the database connection details requires Library(DatabaseConnector)}

\item{cdmDatabaseSchema}{class:character - database schema containing cdm tables}

\item{cohortDatabaseSchema}{class:character - database schema containing cohort}

\item{workDatabaseSchema}{class:character -database you writing tables to}

\item{cohortid}{class:numeric - id of cohort in cohort table}

\item{agegroup}{class:numeric - 1=0-20 year olds; 2=20-40 year olds; 3=40-60 year olds; 4=60-80 year olds;5=80-100 year olds}

\item{gender}{class:numeric - gender concept_id (8507- male; 8532-female)}

\item{type}{class:character - features used by clustering (condition i.e. all condition_concept_ids or group i.e. concept sets),}

\item{groupDef}{class:dataframe - a dataframe containing covariate concept_sets - must have the columns definition and concept_id}

\item{historyStart}{class:numeric days prior to index to start searching person records for features}

\item{historyEnd}{class:numeric days prior to index to stop searching person records for features}

\item{loc}{class:character - directory where the results of the clustering are saved}
}
\value{
clusterData class:clusterData - a list containing:
\item{strata}{an ffdf containing the age/gender/row_id for each person in the cohort}
\item{covariates}{an ffdf containing the covariates for each person in the cohort}
\item{covariateRef}{ an ffdf containing details about the covariates}
\item{metadata}{a list containing details about the data extraction}
}
\description{
This function connects to the CDM and constructs the data used to do the clustering
- this is either condition_concept_ids that are recorded during the defined time period relative
to the cohort start date for each person in the cohort or covariate concept_sets that are
specified by using the 'default' grouping or inputing a dataframe with columns: definition and
concept_id specifiying the concept_ids that make up each covariate definition, see examples below.
}
\examples{

# to extract the males ages between 30 and 45 in cdm_test.dbo.cohort with id 21
# and find whether they have the default concept definitions 1 to 60 days prior
# to cohort start:
dbconnection <- DatabaseConnector::createConnectionDetails(dbms = dbms,server = server,
user = user,password = pw,port = port,schema = cdmDatabaseSchema)

data <- dataExtract(dbconnection, cdmDatabaseSchema='cdm_test.dbo',
                    cohortDatabaseSchema='cdm_test.dbo', cohort_id=21,
                    agegroup = c(30,45), gender=8507,
                    type='group', groupDef='default',
                    historyStart=1,historyEnd=60)

# to extract the cluster data using user specified concept sets:
# where definition 1 contains concept_ids: 101,32011,1 and 63434
#       definition 2 contains concept_ids: 12,13
#       definition 3 contains concept_ids: 450453,21435324,232,3424,4534435 and 3453
groupDef <- data.frame(defintion=c(1,1,1,1,2,2,3,3,3,3,3,3),
                       concept_id =c(c(101,32011,1,63434), c(12,13),
                                     c(450453,21435324,232,3424,4534435,3453))
data <- dataExtract(dbconnection, cdmDatabaseSchema='cdm_test.dbo',
                    cohortDatabaseSchema='cdm_test.dbo', cohort_id=21,
                    agegroup = NULL, gender=NULL,
                    type='group', groupDef=groupDef,
                    historyStart=1,historyEnd=180)

}
\seealso{
DatabaseConnector, OhdsiRTools, SqlRender, ggplot2, reshape2, dplyr, plyr
}
\keyword{OHDSI,}
\keyword{clustering}

